refactor: item관련 type정의 개별 파일로 분리
