+++ src/types/item.ts
@@ -1,0 +1,16 @@
+export type ServerItems = {
+  pointItemId: number
+  pointItemName: string
+  thumbnailUrl: string
+  pointPrice: number
+}
+
+export type ServerItem = {
+  itemName: string
+  description: string
+  thumbnail: string
+  price: number
+  enablePoint: number
+  decreasePoint: number
+  remainPoint: number
+}
